Redesign image capture logic
This commit refactors image capture logic from callback based into decorator
based. The decorator passes to the decoratee a factory for context manager which
can be used to acquire the camera resource. Because the decoratee can control
the scope it needs for the camera, the critical section of the camera resource
can be smaller than when the whole capture callback is wrapped into the critical
section.

This also fixes bug in previous implementation that used per-instance locks for
"protecting" the camera resource. Now the lock is properly shared between still
image and video managers.